Update index.mdx to remove invalid style tag
This is causing this file to fail to compile on the new docs site because JSX required `style`s to be an object mapping and not raw strings. I removed the style attribute entirely because it seems mostly unnecessary? But I'm not 100% sure if it's needed for some other reason, so I am happy add it back using `style={{ margin: 0 }}` if preferred!