Bump cadence version to use updated fxamacker/cbor
Currently, cadence is using a feature branch of fxamacker/cbor
that was based on cbor v2.4.0.

This commit bumps version of fxamacker/cbor to use an
updated branch based on fxamacker/cbor v2.7.0.

The feature/stream-mode branch at fxamacker/cbor is a
backward compatible superset of v2.7.0 that has
extra features used by atree and cadence.